requirements.txt
asgiref==3.9.1
coverage==7.11.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
sqlparse==0.5.3
